Updated Base
- Base updated to version 1.37
- Pets attacking
- Groovy has to be renamed after build

+++ src/games/stendhal/server/maps/deathmatch/DeathmatchEngine.java
@@ -178,7 +178,7 @@ class DeathmatchEngine implements TurnListener {
 		} else {
 			int xp = player.getLevel() * 80;
 			if (xp > player.getXP()) {
-				xp = (int) player.getXP();
+				xp = player.getXP();
 			}
 			player.subXP(xp);
 		}
