+++ src/games/stendhal/server/maps/quests/SheepGrowing.java
@@ -315,7 +315,7 @@ public class SheepGrowing extends AbstractQuest {
             @Override
 			public void fire(Player player, Sentence sentence, EventRaiser npc) {
                 // give XP to level 2
-                long reward = Level.getXP( 2 ) - player.getXP();
+                int reward = Level.getXP( 2 ) - player.getXP();
                 if(reward > MIN_XP_GAIN) {
                     player.addXP(reward);
                 } else {
